backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dilmer <altamiranocd99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "2.3.3"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
pyjwt = "^2.8.0"
flask-migrate = "^4.0.5"
flask-bcrypt = "^1.0.1"
faker = "^20.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
